Added s-expressions
Result can now be a symbol or an s-expression in addition to
integers/decimals and errors. Note: the names for the values have been
changed to be shorter.

The constructors for Result have been modified to now return pointers.

A resultFree function has been added which recursively frees all the
elements of an s-expression

An s-expression is stored internally as an anonymous struct (named list)
which stores the number of cells in the s-expression and a pointer to a
pointer to a Result. The pointer-to-pointer value is intended to be used
as a list of pointers to Results, which will make up the elements of the
list. The struct has been left anonymous on purpose to ensure that
s-expressions/lists are modified only through their enclosing Result
struct.

Error and symbol types have been changed to now both just carry a char*
to the error/symbol information instead of passing around error/symbol
enums

The error enum has been removed but the symbol enum is still used by
parseTag since the parsers raw output is too messy to deal with in
eval() directly. The common error messagese previously denoted using the
error enum are now declared as #defines.

evalOp has been renamed to evalSym and the (temporary) code which
ensures 2-arity has been moved there; this allows eval to always
recursively evaluate anything that it comes across and leaves it upto
evalSym() to reject evaluations based on arity.

The printing functions have been updated to handle the new kinds of
result values. Since error values now carry the error information
themselves, reprError() is no longer needed and has been removed.