Fix memory bug
This shit took me 7 hours to track down for some reason.

The bug didn't even trigger on the normal build 99% of the time; the
only time I got it to trigger on the normal build was by using a massive
input size and even then it happened like 10% of the time.

ASan caught it though.