Add macros to simplify evalSym and bool expressions
ERR_OUT and VAL_OUT free the sexpr and return an error or a value
respectively.

ASSERT_ARITY errors out (by calling ERR_OUT) if the arity is not what we
expect

Also changed the res to be initialized to 1. This does not affect
add/sub/mul since they just overwrite it and it makes the power
calculation cleaner since we dont have to assign an acc anymore.

Note: The power calculation has a bug making it incapable of handling
negative exponents; negative bases are handled fine.

Also added isnot macro --> expands to !=